requirements.txt
Django
pillow
apk_parse_ph4
djangorestframework
markdown
django-filter